Added script for homogenized data format QC function
The "homog.QC()" function uses the "HMG_QC_constraints" sheet from the QC folder on the GDrive to check that homog column data fits the correct format and value range.

Required next step is to complete QC limits in the "HMG_QC_constraints" sheet.

Structure is in place to add almost any QC check we'd like on the homog data files.  

Function takes input of the homog data as dataframe, so we can likely code up a script to run all prior homogenized data on the Gdrive through the function, if useful.

Note: The QC function does not currently break if a condition is failed.  Code is in place to do so, but commented out for current use.