Homog_QC_ftn updated to use the "Key_Key_V2_lookup" sheet
The if statement chains are now a bit complicated, but the function will now attempt to run QC on every var in the lookup sheet and returns messages if values do not fit class, min and max checks. Also returns message is QC value is found in the lookup table. 

Included a check for percent data that is all in range 0-1. Prints warning to console if true.

Next steps:
1) QC for time/date data?

2) Send QC output to log file? Maybe not needed if run within soilHarmonization?

3) What additional QC constraints do we want to put in place?