Updated filter script for control data
Changed filter to step through levels from top (L1, tx_L1) to bottom (L6, tx_L6), filtering the data on the first level column with a match to the control value.

Still need to QC the data output!

Also, I left the previous filter method (keeps data if control value is found in any of the levels columns) in the script, but have it commented out for now. May be useful down the line if we need to create exceptions for individual datasets with wonky control columns/values.